+++ examples/rest_api/mlflow_tracking_rest_api.py
@@ -15,7 +15,7 @@ import pwd
 
 import requests
 
-from mlflow.utils.time_utils import get_current_time_millis
+from mlflow.utils.time import get_current_time_millis
 
 _DEFAULT_USER_ID = "unknown"
