+++ mlflow/models/evaluation/default_evaluator.py
@@ -40,7 +40,7 @@ from mlflow.pyfunc import _ServedPyFuncModel
 from mlflow.sklearn import _SklearnModelWrapper
 from mlflow.utils.file_utils import TempDir
 from mlflow.utils.proto_json_utils import NumpyEncoder
-from mlflow.utils.time_utils import get_current_time_millis
+from mlflow.utils.time import get_current_time_millis
 
 _logger = logging.getLogger(__name__)
