+++ mlflow/runs.py
@@ -9,7 +9,7 @@ from mlflow.entities import ViewType
 from mlflow.environment_variables import MLFLOW_EXPERIMENT_ID
 from mlflow.tracking import _get_store
 from mlflow.utils.string_utils import _create_table
-from mlflow.utils.time_utils import conv_longdate_to_str
+from mlflow.utils.time import conv_longdate_to_str
 
 RUN_ID = click.option("--run-id", type=click.STRING, required=True)
