+++ mlflow/store/model_registry/dbmodels/models.py
@@ -19,7 +19,7 @@ from mlflow.entities.model_registry import (
 from mlflow.entities.model_registry.model_version_stages import STAGE_DELETED_INTERNAL, STAGE_NONE
 from mlflow.entities.model_registry.model_version_status import ModelVersionStatus
 from mlflow.store.db.base_sql_model import Base
-from mlflow.utils.time_utils import get_current_time_millis
+from mlflow.utils.time import get_current_time_millis
 
 
 class SqlRegisteredModel(Base):
