+++ tests/entities/conftest.py
@@ -16,7 +16,7 @@ from mlflow.entities import (
     RunStatus,
     RunTag,
 )
-from mlflow.utils.time_utils import get_current_time_millis
+from mlflow.utils.time import get_current_time_millis
 
 from tests.helper_functions import random_int, random_str
